requirements.txt
pyodbc==4.0.39
Faker==37.3.0
tqdm==4.67.1
python-dotenv==1.0.0
PyYAML==6.0.1
pandas==2.2.0
matplotlib==3.8.2
seaborn==0.13.2
sqlalchemy==2.0.27 